+++ NEWS.md
@@ -7,6 +7,7 @@
 * `recodeGADS()` now allows recoding of unlabeled values
 * `checkEmptyValLabels()` and `checkMissingValLabels()` provide cleaner output
 * `write_spss2()` for writing `.sav` files via a text file and a `SPSS` syntax
+* `import_spss()` and `write_spss()` now deal correctly with missing codes for character variables
 
 # Internal
 * better performance of `applyChangeMeta()` on value level